moved VenueCarde functionality into useVenueCard
